+++ contracts/multisig-prover/src/contract.rs
@@ -131,22 +131,20 @@ mod tests {
     use cosmwasm_std::{
         from_binary,
         testing::{mock_dependencies, mock_env, mock_info, MockApi, MockQuerier, MockStorage},
-        Addr, Empty, Fraction, OwnedDeps, QuerierResult, SubMsgResponse, SubMsgResult, Uint128,
-        Uint256, Uint64, WasmQuery,
+        Addr, Empty, Fraction, OwnedDeps, SubMsgResponse, SubMsgResult, Uint128, Uint256, Uint64,
     };
 
     use axelar_wasm_std::{MajorityThreshold, Threshold, VerificationStatus};
-    use multisig::{
-        msg::{Multisig, Signer},
-        types::MultisigState,
-        worker_set::WorkerSet,
-    };
+    use multisig::{msg::Signer, worker_set::WorkerSet};
     use router_api::CrossChainId;
-    use service_registry::state::{
-        AuthorizationState, BondingState, WeightedWorker, Worker, WORKER_WEIGHT,
-    };
 
-    use crate::contract::execute::should_update_worker_set;
+    use crate::{
+        contract::execute::should_update_worker_set,
+        test::test_utils::{
+            mock_querier_handler, ADMIN, COORDINATOR_ADDRESS, GATEWAY_ADDRESS, GOVERNANCE,
+            MULTISIG_ADDRESS, SERVICE_NAME, SERVICE_REGISTRY_ADDRESS, VOTING_VERIFIER_ADDRESS,
+        },
+    };
     use crate::{
         encoding::Encoder,
         msg::{GetProofResponse, ProofStatus},
@@ -155,16 +153,8 @@ mod tests {
 
     use super::*;
 
-    const GATEWAY_ADDRESS: &str = "gateway";
-    const MULTISIG_ADDRESS: &str = "multisig";
-    const COORDINATOR_ADDRESS: &str = "coordinator";
-    const SERVICE_REGISTRY_ADDRESS: &str = "service_registry";
-    const VOTING_VERIFIER_ADDRESS: &str = "voting_verifier";
-    const ADMIN: &str = "admin";
-    const GOVERNANCE: &str = "governance";
     const RELAYER: &str = "relayer";
     const MULTISIG_SESSION_ID: Uint64 = Uint64::one();
-    const SERVICE_NAME: &str = "validators";
 
     pub fn setup_test_case() -> OwnedDeps<MockStorage, MockApi, MockQuerier, Empty> {
         let mut deps = mock_dependencies();
@@ -199,107 +189,6 @@ mod tests {
         .unwrap();
 
         deps
-    }
-
-    fn mock_querier_handler(
-        operators: Vec<TestOperator>,
-        worker_set_status: VerificationStatus,
-    ) -> impl Fn(&WasmQuery) -> QuerierResult {
-        move |wq: &WasmQuery| match wq {
-            WasmQuery::Smart { contract_addr, .. } if contract_addr == GATEWAY_ADDRESS => {
-                gateway_mock_querier_handler()
-            }
-            WasmQuery::Smart { contract_addr, msg } if contract_addr == MULTISIG_ADDRESS => {
-                multisig_mock_querier_handler(from_binary(msg).unwrap(), operators.clone())
-            }
-            WasmQuery::Smart { contract_addr, .. } if contract_addr == SERVICE_REGISTRY_ADDRESS => {
-                service_registry_mock_querier_handler(operators.clone())
-            }
-            WasmQuery::Smart { contract_addr, .. } if contract_addr == VOTING_VERIFIER_ADDRESS => {
-                voting_verifier_mock_querier_handler(worker_set_status)
-            }
-            _ => panic!("unexpected query: {:?}", wq),
-        }
-    }
-
-    fn gateway_mock_querier_handler() -> QuerierResult {
-        Ok(to_binary(&test_data::messages()).into()).into()
-    }
-
-    fn multisig_mock_querier_handler(
-        msg: multisig::msg::QueryMsg,
-        operators: Vec<TestOperator>,
-    ) -> QuerierResult {
-        let result = match msg {
-            multisig::msg::QueryMsg::GetMultisig { session_id: _ } => {
-                to_binary(&mock_get_multisig(operators))
-            }
-            multisig::msg::QueryMsg::GetPublicKey {
-                worker_address,
-                key_type: _,
-            } => to_binary(
-                &operators
-                    .iter()
-                    .find(|op| op.address == worker_address)
-                    .unwrap()
-                    .pub_key,
-            ),
-            _ => panic!("unexpected query: {:?}", msg),
-        };
-
-        Ok(result.into()).into()
-    }
-
-    fn mock_get_multisig(operators: Vec<TestOperator>) -> Multisig {
-        let quorum = test_data::quorum();
-
-        let signers = operators
-            .into_iter()
-            .map(|op| {
-                (
-                    Signer {
-                        address: op.address,
-                        weight: op.weight,
-                        pub_key: op.pub_key,
-                    },
-                    op.signature,
-                )
-            })
-            .collect::<Vec<_>>();
-
-        Multisig {
-            state: MultisigState::Completed {
-                completed_at: 12345,
-            },
-            quorum,
-            signers,
-        }
-    }
-
-    fn service_registry_mock_querier_handler(operators: Vec<TestOperator>) -> QuerierResult {
-        Ok(to_binary(
-            &operators
-                .clone()
-                .into_iter()
-                .map(|op| WeightedWorker {
-                    worker_info: Worker {
-                        address: op.address,
-                        bonding_state: BondingState::Bonded {
-                            amount: op.weight.try_into().unwrap(),
-                        },
-                        authorization_state: AuthorizationState::Authorized,
-                        service_name: SERVICE_NAME.to_string(),
-                    },
-                    weight: WORKER_WEIGHT,
-                })
-                .collect::<Vec<WeightedWorker>>(),
-        )
-        .into())
-        .into()
-    }
-
-    fn voting_verifier_mock_querier_handler(status: VerificationStatus) -> QuerierResult {
-        Ok(to_binary(&status).into()).into()
     }
 
     fn execute_update_worker_set(
@@ -345,7 +234,7 @@ mod tests {
     }
 
     fn reply_construct_proof(deps: DepsMut) -> Result<Response, axelar_wasm_std::ContractError> {
-        let session_id = to_binary(&Uint64::one()).unwrap();
+        let session_id = to_binary(&MULTISIG_SESSION_ID).unwrap();
 
         let response = SubMsgResponse {
             events: vec![],
