+++ contracts/multisig-prover/src/test/mod.rs
@@ -1,2 +1,4 @@
 #[cfg(test)]
 pub mod test_data;
+#[cfg(test)]
+pub mod test_utils;
